get latest version from release instead of set the default value
